requirements.txt
opencv-python
pillow
pycryptodome
randcrack
requests
tqdm
websockets